Extract getTopViewController
