+++ podcasts/Analytics/Helpers/AnalyticsCoordinator.swift
@@ -59,7 +59,7 @@ class AnalyticsCoordinator {
                 return currentSource
             }
 
-            return (getTopViewController() as? AnalyticsSourceProvider)?.analyticsSource ?? .unknown
+            return (UIApplication.shared.getTopViewController() as? AnalyticsSourceProvider)?.analyticsSource ?? .unknown
         }
 
         func track(_ event: AnalyticsEvent, properties: [String: Any]? = nil) {
@@ -74,21 +74,6 @@ class AnalyticsCoordinator {
             let defaultProperties: [String: Any] = ["source": currentAnalyticsSource]
             let mergedProperties = defaultProperties.merging(properties ?? [:]) { current, _ in current }
             Analytics.track(event, properties: mergedProperties)
-        }
-
-    func getTopViewController(base: UIViewController? = SceneHelper.rootViewController()) -> UIViewController? {
-            guard UIApplication.shared.applicationState == .active else {
-                return nil
-            }
-
-            if let nav = base as? UINavigationController {
-                return getTopViewController(base: nav.visibleViewController)
-            } else if let tab = base as? UITabBarController, let selected = tab.selectedViewController {
-                return getTopViewController(base: selected)
-            } else if let presented = base?.presentedViewController {
-                return getTopViewController(base: presented)
-            }
-            return base
         }
     #else
         /// NOOP track event to preventing needing to wrap all the events in #if checks
